blog_api/requirements.txt
asgiref==3.8.1
dataclasses==0.6
pymongo==3.11.4
python-dotenv==1.0.1
sqlparse==0.2.4
tzdata==2024.2
